requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
distro==1.9.0
faiss-cpu==1.11.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy>=1.26.0,<2.0.0  # PyTorch 2.2 not compatible with NumPy 2.x
openai==1.84.0
packaging==25.0
pandas==2.3.0
pydantic==2.11.5
pydantic_core==2.33.2
PyMuPDF==1.26.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2

# Added for project infrastructure
PyYAML>=6.0
pyarrow>=15.0.0
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.4.0
spacy>=3.7.0
